refactor: separated core logic
