+++ map/autoware_map_loader/CMakeLists.txt
@@ -8,10 +8,15 @@ find_package(PCL REQUIRED COMPONENTS common io filters)
 
 ament_auto_add_library(pointcloud_map_loader_node SHARED
   src/pointcloud_map_loader/pointcloud_map_loader_node.cpp
+  src/pointcloud_map_loader/pointcloud_map_loader.cpp
   src/pointcloud_map_loader/pointcloud_map_loader_module.cpp
+  src/pointcloud_map_loader/partial_map_loader.cpp
   src/pointcloud_map_loader/partial_map_loader_module.cpp
+  src/pointcloud_map_loader/differential_map_loader.cpp
   src/pointcloud_map_loader/differential_map_loader_module.cpp
+  src/pointcloud_map_loader/selected_map_loader.cpp
   src/pointcloud_map_loader/selected_map_loader_module.cpp
+  src/pointcloud_map_loader/pointcloud_map_loader_node_core.cpp
   src/pointcloud_map_loader/utils.cpp
 )
 target_link_libraries(pointcloud_map_loader_node ${PCL_LIBRARIES})
