+++ map/autoware_map_loader/src/pointcloud_map_loader/pointcloud_map_loader_module.hpp
@@ -40,9 +40,6 @@ public:
 private:
   rclcpp::Logger logger_;
   rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr pub_pointcloud_map_;
-
-  [[nodiscard]] sensor_msgs::msg::PointCloud2 load_pcd_files(
-    const std::vector<std::string> & pcd_paths, const boost::optional<float> leaf_size) const;
 };
 }  // namespace autoware::map_loader
